requirements.txt
streamlit
google-generativeai
python-dotenv
openai==0.28